+++ aics-apis/aics-api/src/main/java/kgu/developers/apis/api/file/presentation/exception/FileIsNullException.java
@@ -1,8 +1,8 @@
-package kgu.developers.apis.api.post.presentation.exception;
+package kgu.developers.apis.api.file.presentation.exception;
 
 import kgu.developers.core.common.exception.CustomException;
 
-import static kgu.developers.apis.api.post.presentation.exception.FileExceptionCode.FILE_IS_NULL_EXCEPTION;
+import static kgu.developers.apis.api.file.presentation.exception.FileExceptionCode.FILE_IS_NULL_EXCEPTION;
 
 public class FileIsNullException extends CustomException {
 	public FileIsNullException() {
